Append logs to Manager Spreadsheet
Appending log row is somehow much slower than other operations (updating
cells).
Inserting rows on top is even more slow (~1 minute/operation) so we
opted for appending at the bottom and dealing with ordering in the
spreadsheet itself.

This feature is a very good candidate for being run in a Thread.